refactor: 배치 디렉토리 구조 변경 및 배치 주석
