+++ src/main/java/com/profect/tickle/batch/domain/settlement/dto/SettlementDailyFindTargetDto.java
@@ -1,4 +1,4 @@
-package com.profect.tickle.domain.settlement.dto.batch;
+package com.profect.tickle.batch.domain.settlement.dto;
 
 import lombok.Getter;
