+++ src/main/java/com/profect/tickle/batch/domain/settlement/mapper/SettlementDetailMapper.java
@@ -1,6 +1,6 @@
-package com.profect.tickle.domain.settlement.mapper;
+package com.profect.tickle.batch.domain.settlement.mapper;
 
-import com.profect.tickle.domain.settlement.dto.batch.SettlementDetailFindTargetDto;
+import com.profect.tickle.batch.domain.settlement.dto.SettlementDetailFindTargetDto;
 import com.profect.tickle.domain.settlement.entity.SettlementDetail;
 import com.profect.tickle.global.status.Status;
 import org.apache.ibatis.annotations.Mapper;
