+++ src/main/java/com/profect/tickle/domain/settlement/service/SettlementMonthlyService.java
@@ -1,11 +1,7 @@
 package com.profect.tickle.domain.settlement.service;
 
-import com.profect.tickle.domain.member.entity.Member;
-import com.profect.tickle.domain.member.repository.MemberRepository;
-import com.profect.tickle.domain.settlement.dto.batch.SettlementMonthlyFindTargetDto;
-import com.profect.tickle.domain.settlement.entity.SettlementMonthly;
-import com.profect.tickle.domain.settlement.mapper.SettlementMonthlyMapper;
-import com.profect.tickle.domain.settlement.util.SettlementTimeUtil;
+import com.profect.tickle.batch.domain.settlement.mapper.SettlementMonthlyMapper;
+import com.profect.tickle.batch.domain.settlement.timeUtil.SettlementTimeUtil;
 import com.profect.tickle.global.exception.BusinessException;
 import com.profect.tickle.global.exception.ErrorCode;
 import com.profect.tickle.global.status.Status;
@@ -19,10 +15,6 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.time.Instant;
 import java.time.LocalDate;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -30,53 +22,7 @@ import java.util.Optional;
 public class SettlementMonthlyService {
 
     private final SettlementMonthlyMapper settlementMonthlyMapper;
-    private final MemberRepository memberRepository;
     private final StatusProvider statusProvider;
-
-    public void getSettlementMonthly(){
-        HashMap<String, Object> map = new HashMap<>();
-
-        // 정산 생성 시간
-        Instant settlementDate = Instant.now();
-
-        // 날짜 유틸 yyyy, m, week
-        // 00시00분30초에 어제 날짜 기준으로 해당 주차에 포함되는 주간 정산 데이터 집계
-        LocalDate today = LocalDate.now();
-        SettlementTimeUtil period = SettlementTimeUtil.get(today);
-        map.put("year", period.yearStr());
-        map.put("month", period.monthStr());
-        map.put("now", settlementDate);
-
-        // 월간에 upsert할 주간 정산 조회
-        List<SettlementMonthlyFindTargetDto> aggregates =
-                Optional.ofNullable(settlementMonthlyMapper.aggregateFromWeeklyToMonthly(map))
-                        .orElseThrow(() -> new BusinessException(ErrorCode.SETTLEMENT_TARGET_DB_ERROR));
-
-        if(aggregates.isEmpty()){
-            log.error("정산 대상 데이터가 존재하지 않습니다.");
-        }
-
-        // 월간에 upsert
-        List<SettlementMonthly> monthlyList = new ArrayList<>();
-        for(SettlementMonthlyFindTargetDto dto : aggregates){
-            Member member = memberRepository.findById(dto.getMemberId())
-                    .orElseThrow(() -> new BusinessException(ErrorCode.MEMBER_NOT_FOUND));
-            Status settlementStatus = statusProvider.provide(Settlement.SCHEDULED);
-
-            SettlementMonthly stlMonthly = SettlementMonthly.create(dto, member, settlementStatus);
-            monthlyList.add(stlMonthly);
-        }
-        try {
-            settlementMonthlyMapper.upsertSettlementMonthly(monthlyList);
-        } catch (DataAccessException dae) {
-            log.error("SettlementMonthly upsert 오류, list={}", monthlyList);
-            // 에러 정보 상세 출력
-            log.error("에러 메시지: {}", dae.getMessage());
-            log.error("에러 원인: ", dae.getCause());
-            log.error("스택 트레이스:", dae);
-            throw new BusinessException(ErrorCode.SETTLEMENT_UPSERT_FAILED);
-        }
-    }
 
     /**
      * 오늘이 1일이면 지난 달 정산 내역 업데이트
