+++ src/main/resources/mapper/settlement/SettlementMonthlyMapper.xml
@@ -2,11 +2,11 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.profect.tickle.domain.settlement.mapper.SettlementMonthlyMapper">
+<mapper namespace="com.profect.tickle.batch.domain.settlement.mapper.SettlementMonthlyMapper">
 
     <!-- 주간정산 집계하여 월간정산 -->
     <select id="aggregateFromWeeklyToMonthly" parameterType="map"
-            resultType="com.profect.tickle.domain.settlement.dto.batch.SettlementMonthlyFindTargetDto">
+            resultType="com.profect.tickle.batch.domain.settlement.dto.SettlementMonthlyFindTargetDto">
         SELECT
             member_id AS memberId,
             performance_title AS performanceTitle,
@@ -19,51 +19,13 @@
             SUM(settlement_weekly_commission) AS monthlyCommission,
             SUM(settlement_weekly_net_amount) AS monthlyNetAmount
         FROM settlement_weekly
-        WHERE settlement_year = #{year}
-        AND settlement_month = #{month}
+        WHERE COALESCE(settlement_weekly_updated_at, settlement_weekly_created_at)
+         &gt; #{lastTimeSeconds}
+            AND settlement_year = #{year}
+            AND settlement_month = #{month}
         GROUP BY member_id,
                  performance_title
     </select>
-
-    <!-- 월간 정산 누적 집계 upsert -->
-    <insert id="upsertSettlementMonthly" parameterType="map">
-        INSERT INTO settlement_monthly (
-            member_id,
-            status_id,
-            performance_title,
-            settlement_year,
-            settlement_month,
-            settlement_monthly_sales_amount,
-            settlement_monthly_refund_amount,
-            settlement_monthly_gross_amount,
-            settlement_monthly_commission,
-            settlement_monthly_net_amount,
-            settlement_monthly_created_at
-        ) VALUES
-        <foreach collection="list" item="monthly" separator=",">
-        (
-            #{monthly.member.id},
-            #{monthly.status.id},
-            #{monthly.performanceTitle},
-            #{monthly.year},
-            #{monthly.month},
-            #{monthly.monthlySalesAmount},
-            #{monthly.monthlyRefundAmount},
-            #{monthly.monthlyGrossAmount},
-            #{monthly.monthlyCommission},
-            #{monthly.monthlyNetAmount},
-            #{monthly.monthlyCreatedAt}
-        )
-        </foreach>
-        ON CONFLICT (member_id, performance_title, settlement_year, settlement_month)
-        DO UPDATE SET
-            settlement_monthly_sales_amount = EXCLUDED.settlement_monthly_sales_amount,
-            settlement_monthly_refund_amount = EXCLUDED.settlement_monthly_refund_amount,
-            settlement_monthly_gross_amount = EXCLUDED.settlement_monthly_gross_amount,
-            settlement_monthly_commission = EXCLUDED.settlement_monthly_commission,
-            settlement_monthly_net_amount = EXCLUDED.settlement_monthly_net_amount,
-            settlement_monthly_updated_at = EXCLUDED.settlement_monthly_created_at
-    </insert>
 
     <!-- 1일 기준 이전 달의 정산 상태 업데이트 -->
     <update id="updateSettlementMonthlyStatus">
