+++ src/test/java/com/profect/tickle/batch/domain/settlement/unit/SettlementDetailCsvSerializerTest.java
@@ -2,7 +2,7 @@ package com.profect.tickle.batch.domain.settlement.unit;
 
 import com.profect.tickle.batch.domain.settlement.csvSerializer.SettlementCsvSerializer;
 import com.profect.tickle.domain.member.entity.Member;
-import com.profect.tickle.domain.settlement.dto.batch.SettlementDetailFindTargetDto;
+import com.profect.tickle.batch.domain.settlement.dto.SettlementDetailFindTargetDto;
 import com.profect.tickle.domain.settlement.entity.SettlementDetail;
 import com.profect.tickle.global.status.Status;
 import org.junit.jupiter.api.DisplayName;
@@ -30,7 +30,7 @@ public class SettlementDetailCsvSerializerTest {
         Member member = new Member(1L, "host1@email.com", "pwd1", "host1", null, "",
                 HOST, "01012341234", 0, Instant.now(), Instant.now(), null,
                 "", "", "", "", "",
-                "", "", "", null, null);
+                "", "", "", null);
         Status settlementStatus = Status.create(14L, "정산", (short) 100, "예매 예정", Instant.now());
         SettlementDetail item = SettlementDetail.create(dto, member, settlementStatus, 100000L,
                 0L, 10000L, 1000L, 99000L, Instant.parse("2025-08-31T00:00:00Z"));
